refactoring: rename routers -> routes
